+++ .php-cs-fixer.php
@@ -2,11 +2,10 @@
 
 $finder = PhpCsFixer\Finder::create()
     ->in([
-        __DIR__ . '/app',
-        __DIR__ . '/config',
-        __DIR__ . '/database',
-        __DIR__ . '/routes',
-        __DIR__ . '/tests',
+        __DIR__ . '/scripts',
+        __DIR__ . '/docs',
+        __DIR__ . '/docker',
+        __DIR__ . '/.github',
     ])
     ->name('*.php')
     ->notName('*.blade.php')