+++ vite.config.js
@@ -1,13 +1,8 @@
 import { defineConfig } from 'vite';
-import laravel from 'laravel-vite-plugin';
 import vue from '@vitejs/plugin-vue';
 
 export default defineConfig({
     plugins: [
-        laravel({
-            input: ['resources/css/app.css', 'resources/js/app.js'],
-            refresh: true,
-        }),
         vue({
             template: {
                 transformAssetUrls: {
@@ -19,8 +14,7 @@ export default defineConfig({
     ],
     resolve: {
         alias: {
-            '@': '/resources/js',
-            '~': '/resources',
+            '@': '/src',
         },
     },
     server: {
@@ -28,4 +22,12 @@ export default defineConfig({
             host: 'localhost',
         },
     },
+    build: {
+        // Configuração para template sem assets
+        rollupOptions: {
+            input: {
+                main: './src/main.js'
+            }
+        }
+    }
 });